requirements.txt
pyfingerprint
RPi.GPIO
flask